requirements.txt
streamlit
pandas
numpy
yfinance
altair
pillow
lxml
seaborn
pybase64
matplotlib